+++ src/lib/sup-gui-components/sup/gui/plotting/waveform_helper.cpp
@@ -67,11 +67,12 @@ mvvm::PointItem* GetFirstPoint(const mvvm::LineSeriesDataItem& data_item)
  *
  * First 6 items will get always the same color, then we will use random color.
  */
-std::string GetNextNamedColor(int index)
+std::string GetNextNamedColor(size_t index)
 {
   // Colors from https://www.w3.org/TR/css-color-3/#svg-color
   static std::vector<std::string> colors = {"dodgerblue", "green",   "darkorange",
                                             "crimson",    "fuchsia", "teal"};
+
   return index < colors.size() ? colors[index] : mvvm::utils::RandomNamedColor();
 }
